neither employee-lookup nor job-lookup currently..
provide a working standalone implementation that can pass any kind of
browser testing, should both be rebuilt to have out-of-the-box working
search and show controllers and the border between what they provide and
what the host application must do should be re-evaluated.

(currently the host is expected by employee-lookup to implement handler
to control what is done with employee information after employee lookup
is performed, and in our implementation actually reaches into this gem
when EmployeesController#show builds the params for and interrogates the
model NdJobLookup::HrpyEmployeeJob directly via the #search method, to
retrieve the Jobs data for server-side rendering and display.)

Not happy with this, so it's probably time to cut a release for both
gems (since it works) and build a better usage in the example app.