in the near future these may not be strings
EmployeeJobs model uses a date-type for dates, number type for pidm, and
now personnel-actions is coercing everything into their model classes to
avoid encoding issues to simplify moving queried data from page to page.

(so here, we're taking something that may already be a date, converting
it to a date-string, and then re-parsing it into a date.  not sure what
might break if we just expect a date, and this only changed two lines.)